Django-project/requirements.txt
mysqlclient==2.1.1
django-environ==0.9.0
pika==1.3.1
